Code refactoring and adding of tests
